Fix for compilation on Linux

+++ source/model/telemetry_reader.cpp
@@ -199,7 +199,7 @@ QVariant read_variant(file_reader &reader, telemetry_type type)
 		}
 		case telemetry_type::uint64:
 		{
-			return QVariant(reader.read_uint64());
+			return QVariant((quint64)reader.read_uint64());
 		}
 
 
@@ -209,7 +209,7 @@ QVariant read_variant(file_reader &reader, telemetry_type type)
 		}
 		case telemetry_type::int64:
 		{
-			return QVariant(reader.read_int64());
+			return QVariant((qint64)reader.read_int64());
 		}
 
 		case telemetry_type::floatv:
